APRES TEST ET MODIFICATION

+++ src/service/LogBfManager.java
@@ -5,8 +5,8 @@ import model.LogBuffer;
 public interface LogBfManager {
 	
 	public LogBuffer getLogBuffer();
-	public void increaseSize(int percent);
-	public void decreaseSize(int percent);
+	public void increaseSize(double percent);
+	public void decreaseSize(double percent);
 	public void setSize(int size);
 
 }
